+++ packages/client/auth-core/src/contexts/authlogin.tsx
@@ -1,9 +1,8 @@
 "use client";
 
 import AuthModal from "@/components/AuthModal";
-import { CrossmintService } from "@/services/CrossmintService";
-import { CrossmintEnvironment } from "@/utils";
-import { ReactNode, createContext, useContext, useEffect, useMemo, useState } from "react";
+import { CrossmintServiceFactory } from "@/services/CrossmintService";
+import { type ReactNode, createContext, useContext, useEffect, useMemo, useState } from "react";
 
 type AuthContextType = {
     login: () => void;
@@ -17,19 +16,23 @@ const AuthContext = createContext<AuthContextType>({
     jwt: null,
 });
 
-type AuthProviderParams = {
+export type AuthProviderParams = {
     apiKey: string;
-    environment: CrossmintEnvironment;
     children: ReactNode;
 };
 
-export function AuthProvider({ children, apiKey, environment }: AuthProviderParams) {
-    const [jwtToken, setJwtToken] = useState<string | null>(null);
+const getJwtFromCookie = (): string | null => {
+    if (typeof document === "undefined") {
+        return null; // Check if we're on the client-side
+    }
+    const crossmintSession = document.cookie.split("; ").find((row) => row.startsWith("crossmint-session"));
+    return crossmintSession ? crossmintSession.split("=")[1] : null;
+};
+
+export function AuthProvider({ children, apiKey }: AuthProviderParams) {
+    const [jwtToken, setJwtToken] = useState<string | null>(() => getJwtFromCookie());
     const [modalOpen, setModalOpen] = useState(false);
-    const crossmintService = useMemo(
-        () => new CrossmintService(apiKey, jwtToken, environment),
-        [apiKey, jwtToken, environment]
-    );
+    const crossmintService = useMemo(() => CrossmintServiceFactory.create(apiKey, jwtToken), [apiKey, jwtToken]);
 
     useEffect(() => {
         const crossmintSession = document.cookie.split("; ").find((row) => row.startsWith("crossmint-session"));
@@ -54,7 +57,7 @@ export function AuthProvider({ children, apiKey, environment }: AuthProviderPara
         }
 
         setModalOpen(false);
-    }, [modalOpen, jwtToken]);
+    }, [jwtToken]);
 
     const logout = () => {
         document.cookie = "crossmint-session=; expires=Thu, 01 Jan 1970 00:00:00 UTC; path=/;";
@@ -70,7 +73,7 @@ export function AuthProvider({ children, apiKey, environment }: AuthProviderPara
             //     name: "Test",
             // });
 
-            document.cookie = `crossmint-session=${jwtToken}; path=/;`;
+            document.cookie = `crossmint-session=${jwtToken}; path=/;SameSite=Lax;`;
         }
     }, [jwtToken]);