+++ packages/client/base/src/utils/index.ts
@@ -1,3 +1,16 @@
 export * from "./validate";
 export * from "./ui";
 export * from "./embed";
+export * from "./SDKLogger";
+
+export function isClient() {
+    return typeof window !== "undefined";
+}
+
+export function isLocalhost() {
+    if (process.env.NODE_ENV === "test") {
+        return false;
+    }
+
+    return window.location.origin.includes("localhost");
+}