+++ packages/client/ui/react-ui/src/index.ts
@@ -1,5 +1,7 @@
 export * from "./components";
 export * from "./hooks";
+export * from "./providers";
 
 export { CrossmintEvents, useCrossmintEvents } from "@crossmint/client-sdk-base";
+export { type EVMSmartWallet, Chain } from "@crossmint/client-sdk-smart-wallet";
 export type { CrossmintEvent, CrossmintEventMap } from "@crossmint/client-sdk-base";