+++ packages/client/wallets/smart-wallet/src/SmartWalletSDK.ts
@@ -3,21 +3,27 @@ import { stringify } from "viem";
 import { validateAPIKey } from "@crossmint/common-sdk-base";
 
 import { CrossmintWalletService } from "./api/CrossmintWalletService";
-import { SmartWalletChain } from "./blockchain/chains";
+import type { SmartWalletChain } from "./blockchain/chains";
 import type { EVMSmartWallet } from "./blockchain/wallets";
+import { AccountConfigCache } from "./blockchain/wallets/account/cache";
+import { AccountConfigService } from "./blockchain/wallets/account/config";
+import { AccountCreator } from "./blockchain/wallets/account/creator";
+import { EOACreationStrategy } from "./blockchain/wallets/account/eoa";
+import { PasskeyCreationStrategy } from "./blockchain/wallets/account/passkey";
 import { ClientDecorator } from "./blockchain/wallets/clientDecorator";
 import { SmartWalletService } from "./blockchain/wallets/service";
 import { SmartWalletError } from "./error";
 import { ErrorProcessor } from "./error/processor";
-import { DatadogProvider } from "./services/logging/DatadogProvider";
-import type { SmartWalletSDKInitParams, UserParams, WalletParams } from "./types/Config";
+import { scwDatadogLogger, scwLogger } from "./services";
+import type { SmartWalletSDKInitParams, UserParams, WalletParams } from "./types/params";
+import { SDK_VERSION } from "./utils/constants";
 import { isClient } from "./utils/environment";
-import { logPerformance } from "./utils/log";
 
 export class SmartWalletSDK {
     private constructor(
         private readonly smartWalletService: SmartWalletService,
-        private readonly errorProcessor: ErrorProcessor
+        private readonly errorProcessor: ErrorProcessor,
+        private readonly logger = scwLogger
     ) {}
 
     /**
@@ -35,11 +41,21 @@ export class SmartWalletSDK {
         }
 
         const crossmintService = new CrossmintWalletService(clientApiKey);
-        const errorProcessor = new ErrorProcessor(new DatadogProvider());
-        return new SmartWalletSDK(
-            new SmartWalletService(crossmintService, new ClientDecorator(errorProcessor)),
-            errorProcessor
+        const errorProcessor = new ErrorProcessor(scwDatadogLogger);
+        const accountCreator = new AccountCreator(
+            new EOACreationStrategy(),
+            new PasskeyCreationStrategy(crossmintService.getPasskeyServerUrl(), clientApiKey)
         );
+        const accountCache = new AccountConfigCache(`smart-wallet-${SDK_VERSION}`);
+
+        const smartWalletService = new SmartWalletService(
+            crossmintService,
+            new AccountConfigService(crossmintService, accountCache),
+            accountCreator,
+            new ClientDecorator(errorProcessor)
+        );
+
+        return new SmartWalletSDK(smartWalletService, errorProcessor);
     }
 
     /**
@@ -56,7 +72,7 @@ export class SmartWalletSDK {
         chain: SmartWalletChain,
         walletParams: WalletParams = { signer: { type: "PASSKEY" } }
     ): Promise<EVMSmartWallet> {
-        return logPerformance("GET_OR_CREATE_WALLET", async () => {
+        return this.logger.logPerformance("GET_OR_CREATE_WALLET", async () => {
             try {
                 return await this.smartWalletService.getOrCreate(user, chain, walletParams);
             } catch (error: any) {