+++ packages/client/wallets/smart-wallet/src/utils/constants.ts
@@ -1,11 +1,9 @@
 export const ZERO_DEV_TYPE = "ZeroDev";
 export const CURRENT_VERSION = 0;
-export const DATADOG_CLIENT_TOKEN = "pub035be8a594b35be1887b6ba76c4029ca";
-export const CROSSMINT_DEV_URL = "http://localhost:3000/api";
-export const CROSSMINT_STG_URL = "https://staging.crossmint.com/api";
-export const CROSSMINT_PROD_URL = "https://www.crossmint.com/api";
 export const SCW_SERVICE = "SCW_SDK";
 export const SDK_VERSION = "0.1.0";
 export const API_VERSION = "2024-06-09";
 export const BUNDLER_RPC = "https://rpc.zerodev.app/api/v2/bundler/";
 export const PAYMASTER_RPC = "https://rpc.zerodev.app/api/v2/paymaster/";
+export const SUPPORTED_KERNEL_VERSIONS = ["0.3.1", "0.3.0", "0.2.4"] as const;
+export const SUPPORTED_ENTRYPOINT_VERSIONS = ["v0.6", "v0.7"] as const;