+++ apps/payments/nextjs/CHANGELOG.md
@@ -1,5 +1,64 @@
 # @crossmint/client-sdk-nextjs-starter
 
+## 1.2.11
+
+### Patch Changes
+
+-   @crossmint/client-sdk-react-ui@1.3.8
+
+## 1.2.10
+
+### Patch Changes
+
+-   @crossmint/client-sdk-react-ui@1.3.7
+
+## 1.2.9
+
+### Patch Changes
+
+-   @crossmint/client-sdk-react-ui@1.3.6
+
+## 1.2.8
+
+### Patch Changes
+
+-   @crossmint/client-sdk-react-ui@1.3.5
+
+## 1.2.7
+
+### Patch Changes
+
+-   @crossmint/client-sdk-base@1.2.3
+-   @crossmint/client-sdk-react-ui@1.3.4
+
+## 1.2.6
+
+### Patch Changes
+
+-   @crossmint/client-sdk-react-ui@1.3.3
+
+## 1.2.5
+
+### Patch Changes
+
+-   Updated dependencies [5c60365]
+    -   @crossmint/client-sdk-react-ui@1.3.2
+
+## 1.2.4
+
+### Patch Changes
+
+-   Updated dependencies [12b5eac]
+    -   @crossmint/client-sdk-base@1.2.2
+    -   @crossmint/client-sdk-react-ui@1.3.1
+
+## 1.2.3
+
+### Patch Changes
+
+-   Updated dependencies [71b3778]
+    -   @crossmint/client-sdk-react-ui@1.3.0
+
 ## 1.2.2
 
 ### Patch Changes