+++ apps/payments/vue/CHANGELOG.md
@@ -1,5 +1,50 @@
 # @crossmint/client-sdk-vue-ui-starter
 
+## 1.1.18
+
+### Patch Changes
+
+-   Updated dependencies [a124584]
+    -   @crossmint/client-sdk-smart-wallet@0.1.8
+
+## 1.1.17
+
+### Patch Changes
+
+-   Updated dependencies [691f8f8]
+    -   @crossmint/client-sdk-smart-wallet@0.1.7
+
+## 1.1.16
+
+### Patch Changes
+
+-   @crossmint/client-sdk-base@1.2.3
+-   @crossmint/client-sdk-smart-wallet@0.1.6
+-   @crossmint/client-sdk-vue-ui@1.1.13
+
+## 1.1.15
+
+### Patch Changes
+
+-   Updated dependencies [443372b]
+    -   @crossmint/client-sdk-smart-wallet@0.1.5
+
+## 1.1.14
+
+### Patch Changes
+
+-   Updated dependencies [12b5eac]
+    -   @crossmint/client-sdk-base@1.2.2
+    -   @crossmint/client-sdk-vue-ui@1.1.12
+    -   @crossmint/client-sdk-smart-wallet@0.1.4
+
+## 1.1.13
+
+### Patch Changes
+
+-   Updated dependencies [7bdb67f]
+    -   @crossmint/client-sdk-smart-wallet@0.1.3
+
 ## 1.1.12
 
 ### Patch Changes