+++ apps/wallets/smart-wallet/react/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @crossmint/client-sdk-smart-wallet-react-starter
 
+## 1.5.7
+
+### Patch Changes
+
+-   Updated dependencies [a124584]
+    -   @crossmint/client-sdk-smart-wallet@0.1.8
+
+## 1.5.6
+
+### Patch Changes
+
+-   Updated dependencies [691f8f8]
+    -   @crossmint/client-sdk-smart-wallet@0.1.7
+
+## 1.5.5
+
+### Patch Changes
+
+-   @crossmint/client-sdk-smart-wallet@0.1.6
+
+## 1.5.4
+
+### Patch Changes
+
+-   Updated dependencies [443372b]
+    -   @crossmint/client-sdk-smart-wallet@0.1.5
+
+## 1.5.3
+
+### Patch Changes
+
+-   @crossmint/client-sdk-smart-wallet@0.1.4
+
+## 1.5.2
+
+### Patch Changes
+
+-   Updated dependencies [7bdb67f]
+    -   @crossmint/client-sdk-smart-wallet@0.1.3
+
 ## 1.5.1
 
 ### Patch Changes