+++ packages/client/base/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @crossmint/client-sdk-base
 
+## 1.2.3
+
+### Patch Changes
+
+-   Updated dependencies [8f56ddf]
+    -   @crossmint/common-sdk-base@0.1.2
+
+## 1.2.2
+
+### Patch Changes
+
+-   12b5eac: internal improvements
+-   Updated dependencies [12b5eac]
+    -   @crossmint/common-sdk-base@0.1.1
+
 ## 1.2.1
 
 ### Patch Changes