+++ packages/client/verifiable-credentials/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @crossmint/client-sdk-verifiable-credentials
 
+## 3.4.3
+
+### Patch Changes
+
+-   Updated dependencies [8f56ddf]
+    -   @crossmint/common-sdk-base@0.1.2
+
+## 3.4.2
+
+### Patch Changes
+
+-   Updated dependencies [12b5eac]
+    -   @crossmint/common-sdk-base@0.1.1
+
+## 3.4.1
+
+### Patch Changes
+
+-   ea37a84: Credentials retrieval speedup
+
 ## 3.4.0
 
 ### Minor Changes