+++ packages/client/wallets/smart-wallet/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @crossmint/client-sdk-smart-wallet
 
+## 0.1.8
+
+### Patch Changes
+
+-   a124584: performance monitoring enhancement
+
+## 0.1.7
+
+### Patch Changes
+
+-   691f8f8: passkey registration fix
+
+## 0.1.6
+
+### Patch Changes
+
+-   Updated dependencies [8f56ddf]
+    -   @crossmint/common-sdk-base@0.1.2
+    -   @crossmint/client-sdk-base@1.2.3
+
+## 0.1.5
+
+### Patch Changes
+
+-   443372b: Speed improvements
+
+## 0.1.4
+
+### Patch Changes
+
+-   Updated dependencies [12b5eac]
+    -   @crossmint/client-sdk-base@1.2.2
+    -   @crossmint/common-sdk-base@0.1.1
+
+## 0.1.3
+
+### Patch Changes
+
+-   7bdb67f: Improve errors and logging
+
 ## 0.1.2
 
 ### Patch Changes