+++ packages/common/base/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @crossmint/common-sdk-base
 
+## 0.1.2
+
+### Patch Changes
+
+-   8f56ddf: Make getEnvironmentForKey public
+
+## 0.1.1
+
+### Patch Changes
+
+-   12b5eac: internal improvements
+
 ## 0.1.0
 
 ### Minor Changes